좋아요 수정 및 도수 -> double로 변경

+++ soolkkerbi_project_back/src/main/java/kr/or/skb/product/model/vo/Product.java
@@ -14,7 +14,7 @@ public class Product {
 	private int productNo;
 	private int productCase;
 	private String productName;
-	private int productAlc;
+	private double productAlc;
 	private int productLiter;
 	private String productInfo;
 	private int productPrice;
